Added back Capture, Fainted, and Removed
